--- app/controllers/api/events/v1/decision_reviews_controller.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-class Api::Events::V1::DecisionReviewsController < Api::ApplicationController
-
-  def decision_review_created
-    render json: { message: "Decision Review Created" }, status: :created
-  end
-end
